Restore the fuse API to top level
This restores the fuse programming APIs to the top level and
places the fuse structure definition in the generic example.
